% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Amount_split.R
\name{Amount.split}
\alias{Amount.split}
\title{Amount.split}
\usage{
Amount.split(Amount, K)
}
\arguments{
\item{Amount}{Vector containing the number of inputs and outputs variables of each division and the number of link variables.}

\item{K}{Integer which contains the number of division.}
}
\value{
Return a list
\itemize{
   \item Amount_IN - Vector which contains the number of inputs of each division K.
   \item Amount_OUT - Vector which contains the number of outputs of each division K.
   \item Amount_LINK - Vector which includes the amount of Link variable.
}
}
\description{
This Amount.split function split the Amount vector into the vectors Amount_Input, Amount_Output and Amount_Link
}
\examples{
Amount = c(1,1,1,1,1)
 K = 2
 Amount.split(Amount, K)
}
