% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constraint2_Function.R
\name{Constraint2}
\alias{Constraint2}
\title{Constraint2}
\usage{
Constraint2(
  i_loop,
  direction,
  Input,
  Amount_Input,
  K,
  N,
  sum_m,
  sum_r,
  sum_l,
  Link_obj
)
}
\arguments{
\item{i_loop}{Integer containing the current DMU where the NSBM approach is.}

\item{direction}{String which contains the orientation of the problem (e.g. "non", "input" or "output").}

\item{Input}{Matrix with the data of the input variables.}

\item{Amount_Input}{Vector containing the number of the input variables of each division.}

\item{K}{Integer which contains the number of division.}

\item{N}{Integer which contains the total amount of the DMU's.}

\item{sum_m}{Integer which included the total number of all input variables.}

\item{sum_r}{Integer which included the total number of all output variables.}

\item{sum_l}{Integer which included the total number of all link variables.}

\item{Link_obj}{Flag which is 0 for no link variable in the objective function 1 otherwise.}
}
\value{
Return a matrix which contains the second constraint for the NSBM approach.
}
\description{
The Constraints2 function creates the second constraint for the NSBM approach.
}
\examples{
i_loop = 1
direction = "non"
Input = matrix(c(3, 8,7, 4,2, 3), byrow = TRUE, nrow = 3)
Amount_Input = c(1,1)
K = 2
N = 3
sum_m = 2
sum_r = 2
sum_l = 1
Link_obj = 0
Constraint2(i_loop, direction, Input, Amount_Input, K, N, sum_m, sum_r, sum_l, Link_obj)
}
