% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constraint3_Function.R
\name{Constraint3}
\alias{Constraint3}
\title{Constraint3}
\usage{
Constraint3(
  i_loop,
  direction,
  Output,
  Amount_Output,
  K,
  N,
  sum_m,
  sum_r,
  sum_l,
  Link_obj
)
}
\arguments{
\item{i_loop}{Integer containing the current DMU where the NSBM approach is.}

\item{direction}{String which contains the orientation of the problem (e.g. "non", "input" or "output").}

\item{Output}{Matrix with the data of the output variables.}

\item{Amount_Output}{Vector containing the number of output variables of each division.}

\item{K}{Integer which contains the number of division.}

\item{N}{Integer which contains the total amount of the DMU's.}

\item{sum_m}{Integer which included the total number of all input variables.}

\item{sum_r}{Integer which included the total number of all output variables.}

\item{sum_l}{Integer which included the total number of all link variables.}

\item{Link_obj}{Flag which is 0 for no link variable in the objective function 1 otherwise.}
}
\value{
Return a matrix which contains the third constraint for the NSBM approach.
}
\description{
The Constraint3 function creates the third constraint for the nsbm approach.
}
\examples{
i_loop = 1
direction = "non"
Output = matrix(c(5, 1,1, 5,5, 2), byrow = TRUE, nrow = 3)
Amount_Output = c(1,1)
K = 2
N = 3
sum_m = 2
sum_r = 2
sum_l = 1
Link_obj = 0
Constraint3(i_loop, direction, Output, Amount_Output, K, N, sum_m, sum_r, sum_l, Link_obj)
}
