% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constraint5_Function.R
\name{Constraint5}
\alias{Constraint5}
\title{Constraint5}
\usage{
Constraint5(
  i_loop,
  direction,
  link_con,
  Link,
  Amount_Link,
  K,
  N,
  sum_m,
  sum_r,
  sum_l,
  Link_obj
)
}
\arguments{
\item{i_loop}{Integer containing the current DMU where the NSBM approach is.}

\item{direction}{String which contains the orientation of the problem (e.g. "non", "input" or "output").}

\item{link_con}{Integer which contains the information of the link constraint variant (e.g. 1 for fix and 2 for free constraints).}

\item{Link}{Matrix with the data of the Link variables.}

\item{Amount_Link}{Vector containing the number of link variables between the divisions.}

\item{K}{Integer which contains the number of division.}

\item{N}{Integer which contains the total amount of the DMU's.}

\item{sum_m}{Integer which included the total number of all input variables.}

\item{sum_r}{Integer which included the total number of all output variables.}

\item{sum_l}{Integer which included the total number of all link variables.}

\item{Link_obj}{Flag which is 0 for no link variable in the objective function 1 otherwise.}
}
\value{
Return a matrix with the constraints for the link variables (fix or free variant).
}
\description{
The Constraint5 function creates the constraint for the link variables
}
\examples{
i_loop = 1
direction = "non"
link_con = 1
Link = matrix(c(2,8,1), byrow = TRUE, nrow = 3)
Amount_Link = c(1)
K = 2
N = 3
sum_m = 2
sum_r = 2
sum_l = 1
Link_obj = 0
Constraint5(i_loop, direction, link_con, Link, Amount_Link, K, N, sum_m, sum_r, sum_l, Link_obj)
}
