% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_split.R
\name{data.split}
\alias{data.split}
\title{data.split}
\usage{
data.split(Data, Amount, K, N)
}
\arguments{
\item{Data}{Matrix which contains all the data of the problem.}

\item{Amount}{Vector containing the number of inputs and outputs variables of each division and the number of link variables.}

\item{K}{Integer which contains the number of division.}

\item{N}{Integer which includes the number of DMU's.}
}
\value{
Return a list
\itemize{
    \item In - Matrix with the data of input variables.
    \item Out - Matrix with the data of output variables.
    \item Link - Matrix with the data of the Link variables.
}
}
\description{
The data.split function split the Data matrix into the input, output and link vector
}
\examples{
Data <- matrix(c(3, 5, 8, 1, 2, 7, 1, 4, 5, 8,2, 5, 3, 2, 1),byrow=TRUE, nrow=3)
Amount = c(1,1,1,1,1)
K = 2
N = 3
data.split(Data, Amount, K, N)
}
