% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_all_func.R
\name{load_all_func}
\alias{load_all_func}
\title{load_all_func}
\usage{
load_all_func()
}
\value{
Load all the function in the "Nsbm.function" package.
}
\description{
The load_all_func function load all the function of the nsbm approach.
}
