% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negative_zero_value.R
\name{negative_zero_value}
\alias{negative_zero_value}
\title{negative_zero_values}
\usage{
negative_zero_value(Output_vec, N)
}
\arguments{
\item{Output_vec}{Vector which contain one column of the Output matrix. This vector includes zero/s or negative values.}

\item{N}{Integer which contains the total amount of the DMU's.}
}
\value{
Return the adjusted output vector with only positive values.
}
\description{
The negative_zero_values function adjust negative or zero output values to use them in the nsbm approach calculation. This
function runs only in the non- and output-oriented case.
}
\examples{
Output_vec = matrix(c(3,2,1,0,-1,-2,-3), byrow=TRUE, nrow=7)
N = 7
negative_zero_value(Output_vec, N)
}
