% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsbm_division.R
\name{nsbm.division}
\alias{nsbm.division}
\title{nsbm.division}
\usage{
nsbm.division(
  direction,
  slack_plus,
  slack_minus,
  Input,
  Output,
  Link,
  Amount_Input,
  Amount_Output,
  Amount_Link,
  K,
  N,
  sum_m,
  sum_r,
  Link_obj
)
}
\arguments{
\item{direction}{String which contains the orientation of the problem (e.g. "non", "input" or "output").}

\item{slack_plus}{Matrix which contains the optimal slack plus solution.}

\item{slack_minus}{Matrix which contains the optimal slack minus solution.}

\item{Input}{Matrix with the data of the input variables.}

\item{Output}{Matrix with the data of the output variables.}

\item{Link}{Matrix with the data of the Link variables.}

\item{Amount_Input}{Vector containing the number of input variables of each division.}

\item{Amount_Output}{Vector containing the number of output variables of each division.}

\item{Amount_Link}{Vector containing the number of link variables between the divisions.}

\item{K}{Integer which contains the number of division.}

\item{N}{Integer which contains the total amount of the DMU's.}

\item{sum_m}{Integer which included the total number of all input variables.}

\item{sum_r}{Integer which included the total number of all output variables.}

\item{Link_obj}{Flag which is 0 for no link variable in the objective function 1 otherwise. 1 works only in the non-oriented case.}
}
\value{
Return a vector which contains the results of the division efficiency scores.
}
\description{
The nsbm.division function uses the results of the nsbm approach and calculates the efficiency score of each division.
}
\examples{
direction = "non"
slack_plus <- matrix(c(3,3,3,3,3,3), byrow = TRUE, nrow = 3);
slack_minus <- matrix(c(4,4,4,4,4,4), byrow = TRUE, nrow = 3);
Input <- matrix(c(1,2,1,2,1,2), byrow = TRUE, nrow = 3);
Output <- matrix(c(3,4,3,4,3,4), byrow = TRUE, nrow = 3);
Link <- matrix(c(2,2,2), byrow = TRUE, nrow = 3);
Amount_Input =  c(1,1);
Amount_Output = c(1,1);
Amount_Link = c(1);
K = 2;
N = 3;
sum_m = 2;
sum_r = 2;
Link_obj = 0;
nsbm.division(direction, slack_plus, slack_minus, Input, Output, Link,
Amount_Input, Amount_Output, Amount_Link, K, N, sum_m, sum_r, Link_obj)
}
