% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_NSBM.R
\name{nsbm}
\alias{nsbm}
\title{nsbm}
\usage{
nsbm(
  direction = "non",
  return_to_scale,
  link_con,
  Data,
  Amount,
  K,
  weights,
  NIRS = 0,
  Link_obj
)
}
\arguments{
\item{direction}{String which contains the orientation of the problem. Possible are "non", "input" or "output". Default value is "non".}

\item{return_to_scale}{String which contains the return to scale variant. Possible are "CRS" (constant return to scale) or "VRS" (variant return to scale).}

\item{link_con}{Integer which contains the information of the link constraint variant. Possible are 1 for fix constraint and 2 for free constraint.}

\item{Data}{Matrix which contains all the data of the problem. The data matrix has a special form. First inputs and outputs come from the first division. Then from the second and following division. Finally there are the link variables.}

\item{Amount}{Vector containing the number of inputs and outputs variables of each division and the number of link variables. Start with the number of the division and at the end the number of the link variables.}

\item{K}{Integer which contains the number of division. The number should be an even number.}

\item{weights}{Vector which contains the weights of each division K.}

\item{NIRS}{Flag which is 0 for the normal calculation and 1 for the Non-increasing return to scale (NIRS) calculation. Default value is 0. NIRS is only calculating when return_to_scale = "VRS".}

\item{Link_obj}{Flag which is 0 for no link variable in the objective function 1 otherwise. 1 works only in the non-oriented case.}
}
\value{
Return a list of different results. There are different results for K >1 and K=1
\itemize{
   \item Return for the K>1 case:
      \itemize{
         \item Return for Link_obj = 0 case:
             \itemize{
                \item Eff - Vector which contains the efficiency scores.
                \item divEff - Matrix which contains the division efficiency scores.
                \item weightNSBM - Matrix which contains all the optimal solutions of the variable.
                \item t - Vector which includes the optimal solution of the variable t.
                \item lambda - Matrix which includes the optimal intensity vector lambda.
                \item slacks_plus - Matrix which contains the optimal slack plus solution.
                \item slack_minus - Matrix which contains the optimal slack minus solution.
                \item Input_proj - Matrix which includes the optimal projection of the input variable to the frontier.
                \item Output_proj - Matrix which includes the optimal projection of the output variable to the frontier.
               \item Link_proj - Matrix which includes the optimal projection of the link variable to the frontier.
             }
         \item Return for Link_obj = 1 case:
             \itemize{
                \item Eff - Vector which contains the efficiency scores.
                \item divEff - Matrix which contains the division efficiency scores.
                \item weightNSBM - Matrix which contains all the optimal solutions of the variable.
                \item t - Vector which includes the optimal solution of the variable t.
                \item lambda - Matrix which includes the optimal intensity vector lambda.
                \item slacks_plus - Matrix which contains the optimal slack plus solution.
                \item slack_minus - Matrix which contains the optimal slack minus solution.
            }
       }
   \item Return for the K=1 case:
      \itemize{
         \item Eff - Vector which contains the efficiency scores.
         \item weightNSBM - Matrix which contains all the optimal solutions of the variable.
         \item t - Vector which includes the optimal solution of the variable t.
         \item lambda - Matrix which includes the optimal intensity vector lambda.
         \item slacks_plus - Matrix which contains the optimal slack plus solution
         \item slack_minus - Matrix which contains the optimal slack minus solution.
         \item Input_proj - Matrix which includes the optimal projection of the input variable to the frontier.
         \item Output_proj - Matrix which includes the optimal projection of the output variable to the frontier.
       }
}
}
\description{
This nsbm function Calculates the technical efficiency score by applying the network slacks-based measure (NSBM). It also outputs the
division scores, the optimal variable solution and the projection to the frontier.
}
\examples{
direction = "non"
return_to_scale = "VRS"
link_con = 1
Data <- matrix(c(3, 5, 8, 1, 2, 7, 1, 4, 5, 8,2, 5, 3, 2, 1), byrow=TRUE, nrow=3)
Amount = c(1,1,1,1,1)
K = 2
weights = matrix(c(0.5, 0.5),byrow=TRUE, nrow=1)
NIRS = 0
Link_obj = 0
nsbm(direction, return_to_scale, link_con, Data, Amount, K, weights, NIRS, Link_obj)
}
