% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection_frontier.R
\name{projection.frontier}
\alias{projection.frontier}
\title{projection.frontier}
\usage{
projection.frontier(
  link_con,
  slack_plus,
  slack_minus,
  lambda,
  Input,
  Output,
  Link,
  Amount_Input,
  Amount_Output,
  Amount_Link,
  N,
  K,
  sum_m,
  sum_r,
  sum_l
)
}
\arguments{
\item{link_con}{Integer which contains the information of the link constraint variant (e.g. 1 for fix and 2 for free constraints).}

\item{slack_plus}{Matrix which contains the optimal slack plus solution.}

\item{slack_minus}{Matrix which contains the optimal slack minus solution.}

\item{lambda}{Matrix which contains the optimal solution for the intensity vector.}

\item{Input}{Matrix with the data of the input variables.}

\item{Output}{Matrix with the data of the output variables.}

\item{Link}{Matrix with the data of the link variables.}

\item{Amount_Input}{Vector containing the number of input variables of each division.}

\item{Amount_Output}{Vector containing the number of output variables of each division.}

\item{Amount_Link}{Vector containing the number of link variables between the divisions.}

\item{N}{Integer which contains the total amount of the DMU's.}

\item{K}{Integer which contains the number of division.}

\item{sum_m}{Integer which included the total number of all input variables.}

\item{sum_r}{Integer which included the total number of all output variables.}

\item{sum_l}{Integer which included the total number of all link variables.}
}
\value{
Return a list
 \itemize{
       \item Input_Proj - Optimal projection of the input variable to the frontier.
       \item Output_Proj - Optimal projection of the output variable to the frontier.
       \item Link_Proj - Optimal projection of the link variable to the frontier.
 }
}
\description{
The projection.function calculates the projection for input, output and link variables to the frontier.
}
\examples{
link_con = NaN
 slack_plus <- matrix(c(0.714,0,2.286,0,0,0,0,1,0,0), byrow = TRUE, nrow = 5)
 slack_minus <- matrix(c(0,0.357,0,0.643,0,0,0,0,0,0), byrow = TRUE, nrow = 5)
 lambda <- matrix(c(  0,0,0.3030303,0,0.4848485,
                      0,0,0.4090909,0,0.2727273,
                      0,0,1.0000000,0,0.0000000,
                      0,0,0.6666667,0,0.0000000,
                      0,0,0.0000000,0,1.0000000 ), byrow = TRUE, nrow = 5)
 Input = matrix(c(4,3,6,3,8,1,8,1,2,4), byrow = TRUE, nrow = 5)
 Output = matrix(c(2,3,2,3,6,2,6,1,1,4), byrow = TRUE, nrow = 5)
 Link <- NaN
 Amount_In =  c(2)
 Amount_Out = c(2)
 Amount_L = NaN
 N = 5
 K = 1
 sum_m = 2
 sum_r = 2
 sum_l = 0
 projection.frontier(link_con, slack_plus, slack_minus, lambda, Input, Output,
  Link, Amount_In, Amount_Out, Amount_L, N, K, sum_m, sum_r, sum_l)
}
