% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Righthandside_and_Direction_Function.R
\name{Righthandside_and_Direction}
\alias{Righthandside_and_Direction}
\title{Righthandside_and_Direction}
\usage{
Righthandside_and_Direction(
  i_loop,
  direction,
  return_to_scale,
  link_con,
  NIRS,
  Input,
  Output,
  Link,
  Amount_Link,
  K,
  sum_m,
  sum_r,
  sum_l
)
}
\arguments{
\item{i_loop}{Integer containing the current DMU where the NSBM approach is.}

\item{direction}{String which contains the orientation of the problem (e.g. "non", "input" or "output").}

\item{return_to_scale}{String which contains the return to scale variant (e.g. "CRS" or "VRS").}

\item{link_con}{Integer which contains the information of the link constraint variant (e.g. 1 for fix and 2 for free constraints).}

\item{NIRS}{Flag which is 0 for the normal calculation and 1 for the Non-increasing return to scale (NIRS) calculation.}

\item{Input}{Matrix with the data of the input variables.}

\item{Output}{Matrix with the data of the output variables.}

\item{Link}{Matrix with the data of the link variables.}

\item{Amount_Link}{Vector containing the number of link variables between the divisions.}

\item{K}{Integer which contains the number of division.}

\item{sum_m}{Integer which included the total number of all input variables.}

\item{sum_r}{Integer which included the total number of all output variables.}

\item{sum_l}{Integer which included the total number of all link variables.}
}
\value{
Returns a list
\itemize{
     \item FDIR - Matrix which contains the sign for the nsbm approach.
     \item FRHS - Matrix which contains the right hand side of the nsbm approach.
}
}
\description{
The Righthandside_and_Direction function creates the righthandside (f.rhs) and the direction (f.dir) for the nsbm approach.
}
\examples{
i = 1
direction = "non"
return_to_scale = "VRS"
link_con = 1
NIRS = 0
Input = matrix(c(3, 8,7, 4,2, 3), byrow = TRUE, nrow = 3)
Output = matrix(c(5, 1,1, 5,5, 2), byrow = TRUE, nrow = 3)
Link = matrix(c(2,8,1), byrow = TRUE, nrow = 3)
Amount_L = c(1)
K = 2
sum_m = 2
sum_r = 2
sum_l = 1
Righthandside_and_Direction(i, direction, return_to_scale, link_con, NIRS, Input,
Output, Link, Amount_L, K, sum_m, sum_r, sum_l)
}
