% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slacks_transformation.R
\name{slacks.transformation}
\alias{slacks.transformation}
\title{slacks.transformation}
\usage{
slacks.transformation(
  direction,
  weightsNSBM,
  K,
  N,
  sum_m,
  sum_r,
  sum_l,
  Link_obj
)
}
\arguments{
\item{direction}{String which contains the orientation of the problem (e.g. "non", "input" or "output").}

\item{weightsNSBM}{Matrix which contains all solution of the variables from the NSBM approach.}

\item{K}{Integer which contains the number of division}

\item{N}{Integer which includes the number of DMU's.}

\item{sum_m}{Integer which included the total number of all input variables.}

\item{sum_r}{Integer which included the total number of all output variables.}

\item{sum_l}{Integer which included the total number of all link variables.}

\item{Link_obj}{Flag which is 0 for no link variable in the objective function 1 otherwise. 1 works only in the non-oriented case.}
}
\value{
return a list
 \itemize{
      \item t - Is a vector with the optimal solution of variable t.
      \item lambda - Contains a matrix with the intensity vector lambda.
      \item slack_plus - Contains a matrix with the optimal slack plus solution.
      \item slack_minus - Contains a matrix with the optimal slack minus solution.
 }
}
\description{
The slacks_transformation function uses the the solution of the variables (weightsNSBM)
and decompose them into the individual solution (t, lambda, slack_plus and slack_minus).
}
\examples{
direction = "non"
 weightsNSBM <- matrix(c(0.8,0,0,0.3,0,0.4,0.6,0,0,0.3,0.6,0,0,
 0.4,0,0.2,1.4,0,0,0.4,1,0,0,1,0,0,0,0,0,0,0.6,0,0,0.6,0,0,0,0.6,
 0,0,1,0,0,0,0,1,0,0,0,0), byrow = TRUE, nrow = 5)
 K = 1
 N = 5
 sum_m = 2
 sum_r = 2
 sum_l = 0
 Link_obj = 0
 slacks.transformation(direction, weightsNSBM, K, N, sum_m, sum_r, sum_l, Link_obj)
}
