% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_constraint_Function.R
\name{t_constraint}
\alias{t_constraint}
\title{t_constraint}
\usage{
t_constraint(K, N, sum_r, sum_m, sum_l, Link_obj)
}
\arguments{
\item{K}{Integer which contains the number of division.}

\item{N}{Integer which includes the number of DMU's.}

\item{sum_r}{Integer which included the total number of all output variables.}

\item{sum_m}{Integer which included the total number of all input variables.}

\item{sum_l}{Integer which included the total number of all link variables.}

\item{Link_obj}{Flag which is 0 for no link variable in the objective function 1 otherwise.}
}
\value{
Return a matrix with the constraint that t>0.
}
\description{
The t_constraint function creates a matrix which contains the constraint that t>0.
}
\examples{
K = 2;
N = 3;
sum_r = 2;
sum_m = 2;
sum_l = 1
Link_obj = 0
t_constraint(K, N, sum_r, sum_m, sum_l, Link_obj)
}
