% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VRS_Constraint_Function.R
\name{vrs.constraint}
\alias{vrs.constraint}
\title{vrs.constraint}
\usage{
vrs.constraint(direction, K, N, sum_m, sum_r, sum_l, Link_obj)
}
\arguments{
\item{direction}{String which contains the orientation of the problem (e.g. "non", "input" or "output").}

\item{K}{Integer which contains the number of division.}

\item{N}{Integer which includes the number of DMU's.}

\item{sum_m}{Integer which included the total number of all input variables.}

\item{sum_r}{Integer which included the total number of all output variables.}

\item{sum_l}{Integer which included the total number of all link variables.}

\item{Link_obj}{Flag which is 0 for no link variable in the objective function 1 otherwise.}
}
\value{
Return a matrix with the VRS constraint
}
\description{
The vrs_constraint function creates the VRS constraint for the NBSM approach
}
\examples{
direction = "non"
 K = 2
 N = 3
 sum_m = 2
 sum_r = 2
 sum_l = 1
 Link_obj = 0
 vrs.constraint(direction, K, N, sum_m, sum_r, sum_l, Link_obj)
}
